Migrate to vite
Move and rename files for vite requirements.
Use .default syntax to import cjs from esm.
Remove unnecessary env file.
Edit index.html to remove unnecessary codes.
