+++ package.json
@@ -11,21 +11,23 @@
   },
   "devDependencies": {
     "@dlguswo333/react-simple-spinner": "^0.6.0",
+    "@vitejs/plugin-react": "^4.0.4",
     "electron": "^26.2.2",
     "electron-builder": "^23.6.0",
     "mocha": "^9.2.2",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
-    "react-scripts": "^5.0.1",
-    "sass": "^1.59.3"
+    "sass": "^1.59.3",
+    "vite": "^4.4.9",
+    "vite-plugin-commonjs": "^0.9.0",
+    "vite-plugin-svgr": "^4.0.0"
   },
   "homepage": "./",
   "main": "src/back/main.js",
   "scripts": {
-    "start": "react-scripts start",
-    "build": "react-scripts build",
+    "start": "vite --config src/front/vite.config.js",
+    "build": "vite build --config src/front/vite.config.js",
     "test": "mocha ./src/back/back.test.js",
-    "eject": "react-scripts eject",
     "electron": "electron --trace-warnings .",
     "builder": "npm run build && electron-builder"
   },
