+++ src/front/index.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
 import './index.css';
-import App from './front/App';
+import App from './App';
 
 const reactRoot = ReactDOM.createRoot(document.getElementById('root'));
 reactRoot.render(
